Lots of improvments to TOML config input sanitization
- Unknown keys now result in exception
- Better type restiction for arrays
- All configurable components upgraded to new sanitization scheme

+++ src/positioncombiner/MeanPosition.h
@@ -53,7 +53,7 @@ private:
     bool generate_heading;
     
     // SOURCE position to be used ad an anchor when calculating headings
-    int heading_anchor_idx;
+    int64_t heading_anchor_idx;
 
 };
